Item reservation and removal
When a user accepts a trade, all items in the trade are marked as
"Reserved".  Reserved items remain visible in other trades, but are
marked as reserved, and cannot be part of another trade: attempting to
accept or submit a trade with reserved items is not allowed.  When the
trade is confirmed (both traders confirm the trade) it is marked as
removed and no longer appears in trader's lists of Haves.  Reserved and
removed items do not appear in item searches.